+++ operator-common/src/main/java/io/strimzi/operator/common/model/Ca.java
@@ -508,8 +508,6 @@ public abstract class Ca {
      * @param additionalLabels              The additional labels of the {@code Secrets} created.
      * @param additionalAnnotations         The additional annotations of the {@code Secrets} created.
      * @param ownerRef                      The owner of the {@code Secrets} created.
-     * @param existingServerSecrets         List of existing Secrets with certificates signed by this CA. This is used
-     *                                      to compare the CA generation from their annotations with the CA generation.
      * @param maintenanceWindowSatisfied    Flag indicating whether we are in the maintenance window
      */
     public void createRenewOrReplace(
@@ -518,7 +516,6 @@ public abstract class Ca {
             Map<String, String> additionalLabels,
             Map<String, String> additionalAnnotations,
             OwnerReference ownerRef,
-            List<HasMetadata> existingServerSecrets,
             boolean maintenanceWindowSatisfied
     ) {
         X509Certificate currentCert = cert(caCertSecret, CA_CRT);
@@ -530,7 +527,7 @@ public abstract class Ca {
         if (!generateCa) {
             certData = caCertSecret.getData();
             keyData = singletonMap(CA_KEY, caKeySecret.getData().get(CA_KEY));
-            renewalType = hasCaCertGenerationChanged(existingServerSecrets) ? RenewalType.REPLACE_KEY : RenewalType.NOOP;
+            renewalType = RenewalType.NOOP; // User is managing CA
             caCertsRemoved = false;
         } else {
             this.renewalType = shouldCreateOrRenew(currentCert, maintenanceWindowSatisfied);
@@ -734,7 +731,7 @@ public abstract class Ca {
     }
 
     /**
-     * True if the last call to {@link #createRenewOrReplace(String, Map, Map, Map, OwnerReference, List, boolean)}
+     * True if the last call to {@link #createRenewOrReplace(String, Map, Map, Map, OwnerReference, boolean)}
      * resulted in expired certificates being removed from the CA {@code Secret}.
      * @return Whether any expired certificates were removed.
      */
@@ -743,7 +740,7 @@ public abstract class Ca {
     }
 
     /**
-     * True if the last call to {@link #createRenewOrReplace(String, Map, Map, Map, OwnerReference, List, boolean)}
+     * True if the last call to {@link #createRenewOrReplace(String, Map, Map, Map, OwnerReference, boolean)}
      * resulted in a renewed CA certificate.
      * @return Whether the certificate was renewed.
      */
@@ -752,7 +749,7 @@ public abstract class Ca {
     }
 
     /**
-     * True if the last call to {@link #createRenewOrReplace(String, Map, Map, Map, OwnerReference, List, boolean)}
+     * True if the last call to {@link #createRenewOrReplace(String, Map, Map, Map, OwnerReference, boolean)}
      * resulted in a replaced CA key.
      * @return Whether the key was replaced.
      */