+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Allow using custom certificates for communication with container registry in Kafka Connect Build with Kaniko 
 * Only roll pods once for ClientsCa cert renewal.
   This change also means the restart reason ClientCaCertKeyReplaced is removed and either CaCertRenewed or CaCertHasOldGeneration will be used.
+* Allow rolling update for new cluster CA trust (during Cluster CA key replacement) to continue where it left off before interruption without rolling all pods again.
 
 ### Major changes, deprecations and removals